Added some more text, and a little more.
Paragraphs on game development and other stuff. Also fixed a very specific problem with my footer.

The footer is set up so that I have an Anchor tag to bring you to the top of the page. It is set inside of a paragraph tag, so that it will not be on the same row as the gif at the bottom of the page. Unfortunately, the paragraph tag extends to the edge of the div, so if you hover anywhere on the row, the Hover effect is applied. I fixed this by getting rid of the paragraph tag, and adding a break before the gif, but after the anchor.