Updated Photo
Picture was made smaller to shorten long loading times. This would cause an error where the page would display a large block until you clicked on the "block." Hopefully this will fix the problem. I'll be testing this on the following browsers:

Safari for Macintosh.
Firefox for Windows and Macintosh.
Chrome for Windows and Macintosh.
Edge for Windows.