The GAME Update
I added a little game to my website. At the top of the Index file, you shall find that there are Five hidden keys throughout the website. They are all Capital F's, and turn white when hovered over. If you can find them all... You found them all??

This really wasn't a serious commit, but I kept it because it was a nice exercise to demonstrate CSS Animations, and it fits the tone of my website.