Game Updates, and Small Changes
I felt like the game was too much of a needle in the haystack, so I made it a little easier. 

Now, the Keys (F), will expand in size when hovered, and turn colors immediately. This makes it much easier to spot them.
There are hints throughout various pages, as rewards for reading through my mindless rambles.