Spring Break Update!
Changed a couple things, but the big one:

Changed the illustrator page to "Photos" because I'm including photos in there. Huh, wonder why!

Internally, it is still referred to as "illustrator.html". Whether that will be updated or not is up to how lazy I am about it.

Photos will be added in a separate commit.

If you're wondering why the style on the pictures has it rotated,  it's because my iPhone took the pictures upside down somehow, oh well.